requirements_analyzer/requirements.txt
dataclasses>=0.6
typing-extensions>=4.0.0
pathlib>=1.0.1
python-dotenv>=0.19.0
openai>=1.0.0
markdown>=3.3.0 